third_party/blink/web_tests/external/wpt/html/rendering/replaced-elements/attributes-for-embedded-content-and-images/video-aspect-ratio-expected.txt

This is a testharness.js-based test.
[FAIL] Video width and height attributes are used to infer aspect-ratio
  assert_approx_equals: expected 2.5 +/- 0.001 but got 2
Harness: the test ran to completion.
